2006-08-30	Joel Sherrill <joel@OARcorp.com>
	* ftpd/ftpd.c, libcsupport/include/sys/ioccom.h,
	libcsupport/src/ioctl.c, libnetworking/kern/uipc_mbuf.c,
	libnetworking/libc/inet_addr.c, libnetworking/net/ethernet.h,
	libnetworking/net/if.c, libnetworking/net/if_ethersubr.c,
	libnetworking/net/if_ppp.c, libnetworking/net/if_pppvar.h,
	libnetworking/net/if_var.h, libnetworking/net/ppp_tty.c,
	libnetworking/netinet/in_cksum.c, libnetworking/nfs/bootp_subr.c,
	libnetworking/rtems/rtems_bsdnet_internal.h,
	libnetworking/sys/mbuf.h, pppd/md4.c, pppd/pppd.h: Remove warnings
	due to improper use of int which shows up on 16 bit targets. Added
	ioctl_command_t since IOCTL command argument does not reliably fit
	into 16 bits.

+++ cpukit/libnetworking/sys/mbuf.h
@@ -395,7 +395,7 @@ union mcluster {
 }
 
 /* Length to m_copy to copy all. */
-#define	M_COPYALL	1000000000
+#define	M_COPYALL	(uint32_t)1000000000L
 
 /* Compatibility with 4.3. */
 #define  m_copy(m, o, l)	m_copym((m), (o), (l), M_DONTWAIT)
@@ -413,7 +413,7 @@ extern int	max_protohdr;		/* largest protocol header */
 extern int	max_hdr;		/* largest link+protocol header */
 extern int	max_datalen;		/* MHLEN - max_hdr */
 
-struct	mbuf *m_copym(struct mbuf *, int, int, int);
+struct	mbuf *m_copym(struct mbuf *, int, uint32_t, int);
 struct	mbuf *m_copypacket(struct mbuf *, int);
 struct	mbuf *m_devget(char *, int, int, struct ifnet *,
 			    void (*copy)(char *, caddr_t, u_int));
